Make Fsm abstract, Add SimpleFsm for tests
